# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.GetAllLecturersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Lecturer'
        type: array
      status:
        type: string
    type: object
...
  model.GetAllStudentsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Student'
        type: array
      status:
        type: string
    type: object
...
  model.GetProfileResponse:
    properties:
      data:
        properties:
          email:
            type: string
          full_name:
            type: string
          permissions:
            items:
              type: string
            type: array
          role:
            type: string
          role_id:
            type: string
          user_id:
            type: string
          username:
            type: string
        type: object
      status:
        type: string
    type: object
...
  model.Lecturer:
    properties:
      created_at:
        type: string
      department:
        type: string
      full_name:
        type: string
      id:
        type: string
      lecturer_id:
        type: string
      user_id:
        type: string
    type: object
  model.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.LoginResponse:
    properties:
      data:
        properties:
          refreshToken:
            type: string
          token:
            type: string
          user:
            $ref: '#/definitions/model.LoginUserResponse'
        type: object
      status:
        type: string
    type: object
  model.LoginUserResponse:
    properties:
      fullName:
        type: string
      id:
        type: string
      permissions:
        items:
          type: string
        type: array
      role:
        type: string
      username:
        type: string
    type: object
...
  model.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  model.RefreshTokenResponse:
    properties:
      data:
        properties:
          refreshToken:
            type: string
          token:
            type: string
        type: object
      status:
        type: string
    type: object
...
  model.Student:
    properties:
      academic_year:
        type: string
      advisor_id:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      program_study:
        type: string
      student_id:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: API untuk sistem pelaporan prestasi mahasiswa dengan dukungan multi-role
    (Admin, Dosen Wali, Mahasiswa) dan field prestasi yang fleksibel
  title: Sistem Pelaporan Prestasi Mahasiswa API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login menggunakan username/email dan password untuk mendapatkan
        JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user dari sistem
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - Authentication
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Mengambil profil user yang sedang login beserta role dan permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token menggunakan refresh token yang valid
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Authentication
...
  /lecturers:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua lecturer dari database. Hanya dapat diakses
        oleh admin dengan permission user:manage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetAllLecturersResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all lecturers
      tags:
      - Lecturers
  /lecturers/{id}/advisees:
    get:
      consumes:
      - application/json
      description: Mengambil daftar mahasiswa bimbingan dari lecturer tertentu. Hanya
        dapat diakses oleh admin dengan permission user:manage
      parameters:
      - description: Lecturer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetAllStudentsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get lecturer advisees
      tags:
      - Lecturers
...
  /reports/lecturer:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi mahasiswa bimbingan untuk lecturer yang
        sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get current lecturer report
      tags:
      - Reports
  /reports/lecturer/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil laporan prestasi mahasiswa bimbingan untuk lecturer berdasarkan
        ID
      parameters:
      - description: Lecturer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get lecturer report by ID
      tags:
      - Reports
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"